Remove unnecessary casting

+++ 2019/asteroid/slope.go
@@ -10,8 +10,8 @@ type slope struct {
 }
 
 func calculateSlope(a math.Point, b math.Point) slope {
-	dx := int64(a.X - b.X)
-	dy := int64(a.Y - b.Y)
+	dx := a.X - b.X
+	dy := a.Y - b.Y
 	gcd := math.GCD(math.Abs(dx), math.Abs(dy))
 	return slope{
 		dx / gcd,
